Clean interrupt up a bit
Now APIC is mandatory for RedLeaf

+++ src/tls.rs
@@ -5,7 +5,7 @@ pub const KERNEL_PERCPU_OFFSET: usize = 0x400000;
 pub const KERNEL_PERCPU_SIZE: usize = 0x100000;
 
 /// Copy tdata, clear tbss, set TCB self pointer
-pub unsafe fn init_tcb(cpu_id: usize) -> usize {
+pub unsafe fn init_tcb(cpu_id: u32) -> usize {
     extern {
         /// The starting byte of the thread data segment
         static mut __tdata_start: u8;
@@ -22,7 +22,7 @@ pub unsafe fn init_tcb(cpu_id: usize) -> usize {
         let size = & __tbss_end as *const _ as usize - & __tdata_start as *const _ as usize;
         let tbss_offset = & __tbss_start as *const _ as usize - & __tdata_start as *const _ as usize;
 
-        let start = KERNEL_PERCPU_OFFSET + KERNEL_PERCPU_SIZE * cpu_id;
+        let start = KERNEL_PERCPU_OFFSET + KERNEL_PERCPU_SIZE * (cpu_id as usize);
         let end = start + size;
         tcb_offset = end - mem::size_of::<usize>();
